requirements.txt
flask-jsontools==0.1.7
Flask-SQLAlchemy==3.0.5
future==0.18.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql==0.0.3
mysqlclient==2.2.0
PyMySQL==1.1.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
Werkzeug==2.3.6